+++ src/utils.rs
@@ -7,10 +7,10 @@
 /// # Examples
 /// ```
 /// use vectorize::utils::delta_coord;
-/// let (x1, x2) = (20, 10);
+/// let (x1, x2) = (20., 10.);
 /// delta_coord(x2, x1);
 ///```
 ///
-pub fn delta_coord(x2: isize, x1: isize) -> isize {
-  x2 - x1
+pub fn delta_coord(x2: f64, x1: f64) -> f64 {
+    x2 - x1
 }
